+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 0.11.0 (xxxx-xx-xx)
 
 * Add support for TMC alternative frequencies (tuning info variant 6)
+* Add support for TMC gap parameter and enhanced mode (3A group variant 1)
 * Change PLL bandwidth and lowpass cutoff frequency to improve noise performance
 * Change resampler anti-alias cutoff frequency to allow for lower sample rates
 * Fix missing zero-padding in PIN time string